moving stateful logic back to index to speed things up and provide consistent loading of config, settings, and data
